+++ example/example.dart
@@ -1,7 +1,6 @@
 Future<void> main() async {
-  print(
-      'No example here since this is package is a generic implementation that '
+  print('No example here since this is package is a base implementation that '
       'depends on a backing SQL database to achieve its functionality.\n\n'
-      'Check https://github.com/cachapa/sqlite_crdt or '
+      'See https://github.com/cachapa/sqlite_crdt or '
       'https://github.com/cachapa/postgres_crdt for actual implementations.');
 }