+++ test/sql_util_test.dart
@@ -1,3 +1,4 @@
+import 'package:sql_crdt/sql_crdt.dart';
 import 'package:sql_crdt/src/sql_util.dart';
 import 'package:test/test.dart';
 
@@ -51,6 +52,23 @@ void main() {
         INTERSECT
         SELECT * FROM table6
       '''), {'table1', 'table2', 'table3', 'table4', 'table5', 'table6'});
+    });
+  });
+
+  group('Add changeset clauses', () {
+    test('Simple query', () {
+      final sql = SqlUtil.addChangesetClauses('test', 'SELECT * FROM test',
+          exceptNodeId: 'node_id', modifiedAfter: Hlc.zero('node_id'));
+      expect(sql,
+          "SELECT * FROM test WHERE node_id != 'node_id' AND modified > '1970-01-01T00:00:00.000Z-0000-node_id'");
+    });
+
+    test('Simple query with where clause', () {
+      final sql = SqlUtil.addChangesetClauses(
+          'test', 'SELECT * FROM test WHERE a != ?1 and b = ?2',
+          exceptNodeId: 'node_id', modifiedAfter: Hlc.zero('node_id'));
+      expect(sql,
+          "SELECT * FROM test WHERE node_id != 'node_id' AND modified > '1970-01-01T00:00:00.000Z-0000-node_id' AND a != ?1 AND b = ?2");
     });
   });
 }